moved lib/ to degenotate_lib/ for bioconda recipe
