+++ build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cp degenotate.py ${PREFIX}/bin/.
+mkdir -p ${SP_DIR}
+cp -R degenotate_lib ${SP_DIR}/.
