+++ degenotate_lib/codon_table-old.csv
