+++ degenotate_lib/core.py
